Move Core project to src folder
