--- CSharp/VirtoCommerce.Module.Template/Core/ModuleConstants.cs
@@ -1,15 +1,0 @@
-namespace $safeprojectname$
-{
-    public static class ModuleConstants
-    {
-        public static class Security
-        {
-            public static class Permissions
-            {
-                public const string Read = "$ext_safeprojectname$:read";
-                public const string Update = "$ext_safeprojectname$:update";
-                public const string Delete = "$ext_safeprojectname$:delete";
-            }
-        }
-    }
-}
