--- CSharp/VirtoCommerce.Module.Template/Core/_packages.config
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net461" />
-  <package id="VirtoCommerce.Domain" version="2.25.32" targetFramework="net461" />
-  <package id="VirtoCommerce.Platform.Core" version="2.13.56" targetFramework="net461" />
-</packages>
