+++ CSharp/VirtoCommerce.Module.Template/VirtoCommerce.Module.Template.csproj
@@ -79,8 +79,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="Core\_packages.config" />
     <None Include="Data\_packages.config" />
+    <None Include="src\Core\Core.csproj" />
     <None Include="Tests\_packages.config" />
     <None Include="Web\Readme.md" />
     <None Include="Web\Localizations\en.module.json" />
@@ -91,11 +91,10 @@
     <None Include="Web\Web.config" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="src\Core\ModuleConstants.cs" />
     <None Include="Web\AssemblyInfo.cs" />
     <None Include="Tests\AssemblyInfo.cs" />
-    <None Include="Core\AssemblyInfo.cs" />
     <None Include="Data\AssemblyInfo.cs" />
-    <None Include="Core\ModuleConstants.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -116,10 +115,9 @@
     </BootstrapperPackage>
   </ItemGroup>
   <ItemGroup>
-    <Content Include="Core\Core.csproj" />
-    <Content Include="Core\Core.vstemplate" />
     <Content Include="Data\Data.csproj" />
     <Content Include="Data\Data.vstemplate" />
+    <Content Include="src\Core\Core.vstemplate" />
     <Content Include="Tests\Tests.csproj" />
     <Content Include="Tests\Tests.vstemplate" />
     <Content Include="Web\Content\css\stylesheet.css" />
@@ -138,6 +136,12 @@
     <None Include="Tests\Test.cs" />
     <None Include="Web\Controllers\Api\MyModuleController.cs" />
     <None Include="Web\Module.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="src\Core\Events\" />
+    <Folder Include="src\Core\Models\" />
+    <Folder Include="src\Core\Notifications\" />
+    <Folder Include="src\Core\Services\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
