+++ CSharp/VirtoCommerce.Module.Template/src/Core/Core.vstemplate
@@ -1,6 +1,6 @@
 <VSTemplate Version="3.0.0" xmlns="http://schemas.microsoft.com/developer/vstemplate/2005" Type="Project">
   <TemplateData>
-    <Name>Virto Commerce 2.x Module Core Project</Name>
+    <Name>Virto Commerce 3.x Module Core Project</Name>
     <ProjectType>CSharp</ProjectType>
     <ProjectSubType></ProjectSubType>
     <CreateNewFolder>true</CreateNewFolder>
@@ -9,12 +9,12 @@
   </TemplateData>
   <TemplateContent>
     <Project TargetFileName="$safeprojectname$.csproj" File="Core.csproj" ReplaceParameters="true">
-      <ProjectItem ReplaceParameters="true" TargetFileName="packages.config">_packages.config</ProjectItem>
-      <ProjectItem ReplaceParameters="true" TargetFileName="Properties\AssemblyInfo.cs">AssemblyInfo.cs</ProjectItem>
       <ProjectItem ReplaceParameters="true">ModuleConstants.cs</ProjectItem>
-      
-      <Folder Name="Models" />
-      <Folder Name="Services" />
+        <Folder Name="Events" TargetFolderName="Events" />
+        <Folder Name="Models" TargetFolderName="Models" />
+        <Folder Name="Notifications" TargetFolderName="Notifications" />
+        <Folder Name="Services" TargetFolderName="Services" />
+        <ProjectItem ReplaceParameters="true" TargetFileName="ModuleConstants.cs">ModuleConstants.cs</ProjectItem>
     </Project>
   </TemplateContent>
   <WizardExtension>
