+++ CSharp/VirtoCommerce.Module.Template/src/Core/__TemplateIcon.ico
